deps: upgrade all crates to latest (2026-05)
- regex 1.10→1.12, roxmltree 0.20→0.21, serde 1.0→1.0.228, serde_json 1.0→1.0.149
- clap 4→4.6.1, tokio 1→1.52.3, uuid 1→1.23.1, libc 0.2→0.2.186
- sha2 0.10→0.11: adapt Digest output from LowerHex to iter+format hex
- rusqlite 0.32→0.39: cast priority usize→i64 (ToSql removed for usize)
- axum 0.7→0.8: route path syntax :param→{param}
- reqwest: kept at 0.12.28 (0.13 renamed rustls-tls feature, needs migration)
- fix: temp_sqlite_path now cleans stale files before use (test isolation)
- 52 tests pass, clippy clean, fmt clean

+++ src/storage.rs
@@ -221,7 +221,7 @@ fn insert_intervention_candidates(
             "INSERT INTO intervention_candidates (run_id, priority, event_id, target, intervention_type, reason, expected_effect) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
             params![
                 run_id,
-                candidate.priority,
+                candidate.priority as i64,
                 candidate.event_id,
                 candidate.target,
                 candidate.intervention_type,
